#minor: version serializers, comparators & reporters
