+++ java-snapshot-testing-core/src/main/java/au/com/origin/snapshots/serializers/ToStringSnapshotSerializer.java
@@ -1,8 +1,6 @@
 package au.com.origin.snapshots.serializers;
 
-import au.com.origin.snapshots.Snapshot;
-import au.com.origin.snapshots.SnapshotFile;
-import au.com.origin.snapshots.SnapshotSerializerContext;
+import au.com.origin.snapshots.logging.LoggingHelper;
 import lombok.extern.slf4j.Slf4j;
 
 /**
@@ -11,21 +9,13 @@ import lombok.extern.slf4j.Slf4j;
  * <p>Will render each toString() on a separate line
  */
 @Slf4j
-public class ToStringSnapshotSerializer implements SnapshotSerializer {
-
-  @Override
-  public Snapshot apply(Object object, SnapshotSerializerContext gen) {
-    String body = "[\n" + object.toString() + "\n]";
-    if (body.contains(SnapshotFile.SPLIT_STRING)) {
-      log.warn(
-          "Found 3 consecutive lines in your snapshot \\n\\n\\n. This sequence is reserved as the snapshot separator - replacing with \\n.\\n.\\n");
-      body = body.replaceAll(SnapshotFile.SPLIT_STRING, "\n.\n.\n");
-    }
-    return gen.toSnapshot(body);
-  }
-
-  @Override
-  public String getOutputFormat() {
-    return SerializerType.TEXT.name();
+@Deprecated
+public class ToStringSnapshotSerializer
+    extends au.com.origin.snapshots.serializers.v1.ToStringSnapshotSerializer {
+  public ToStringSnapshotSerializer() {
+    super();
+    LoggingHelper.deprecatedV5(
+        log,
+        "Update to `au.com.origin.snapshots.serializers.v1.ToStringSnapshotSerializer` in `snapshot.properties`");
   }
 }
