+++ java-snapshot-testing-junit5/src/test/resources/snapshot.properties
@@ -1,7 +1,7 @@
-serializer=au.com.origin.snapshots.serializers.ToStringSnapshotSerializer
-serializer.json=au.com.origin.snapshots.jackson.serializers.DeterministicJacksonSnapshotSerializer
-comparator=au.com.origin.snapshots.comparators.PlainTextEqualsComparator
-reporters=au.com.origin.snapshots.reporters.PlainTextSnapshotReporter
+serializer=au.com.origin.snapshots.serializers.v1.ToStringSnapshotSerializer
+serializer.json=au.com.origin.snapshots.jackson.serializers.v1.DeterministicJacksonSnapshotSerializer
+comparator=au.com.origin.snapshots.comparators.v1.PlainTextEqualsComparator
+reporters=au.com.origin.snapshots.reporters.v1.PlainTextSnapshotReporter
 snapshot-dir=__snapshots__
 output-dir=src/test/java
 ci-env-var=CI
