+++ java-snapshot-testing-plugin-jackson/src/main/java/au/com/origin/snapshots/jackson/serializers/DeterministicJacksonSnapshotSerializer.java
@@ -1,7 +1,7 @@
 package au.com.origin.snapshots.jackson.serializers;
 
-import com.fasterxml.jackson.databind.MapperFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import au.com.origin.snapshots.logging.LoggingHelper;
+import lombok.extern.slf4j.Slf4j;
 
 /**
  * Attempts to deterministically render a snapshot.
@@ -11,11 +11,14 @@ import com.fasterxml.jackson.databind.ObjectMapper;
  *
  * <p>Note that collections will be ordered which mar or may not be desirable given your use case.
  */
-public class DeterministicJacksonSnapshotSerializer extends JacksonSnapshotSerializer {
-
-  @Override
-  public void configure(ObjectMapper objectMapper) {
-    objectMapper.enable(MapperFeature.SORT_PROPERTIES_ALPHABETICALLY);
-    objectMapper.registerModule(new DeterministicCollectionModule());
+@Deprecated
+@Slf4j
+public class DeterministicJacksonSnapshotSerializer
+    extends au.com.origin.snapshots.jackson.serializers.v1.DeterministicJacksonSnapshotSerializer {
+  public DeterministicJacksonSnapshotSerializer() {
+    super();
+    LoggingHelper.deprecatedV5(
+        log,
+        "Update to `au.com.origin.snapshots.jackson.serializers.v1.DeterministicJacksonSnapshotSerializer` in `snapshot.properties`");
   }
 }
