+++ java-snapshot-testing-spock/src/test/resources/snapshot.properties
@@ -1,6 +1,6 @@
-serializer=au.com.origin.snapshots.serializers.ToStringSnapshotSerializer
-comparator=au.com.origin.snapshots.comparators.PlainTextEqualsComparator
-reporters=au.com.origin.snapshots.reporters.PlainTextSnapshotReporter
+serializer=au.com.origin.snapshots.serializers.v1.ToStringSnapshotSerializer
+comparator=au.com.origin.snapshots.comparators.v1.PlainTextEqualsComparator
+reporters=au.com.origin.snapshots.reporters.v1.PlainTextSnapshotReporter
 snapshot-dir=__snapshots__
 output-dir=src/test/groovy
 ci-env-var=CI
